Crate proj_crypto

Proj_crypto library crate.

This library contains the cryptography used in my third year project at university. Please do not use this for anything important. This cryptography has not been reviewed.

See the code example in the asymmetric module for a brief overview of what this library can do.

This project is licenced under the terms of the GNU General Public Licence as published by the Free Software Foundation, either version 3 of the licence, or (at your option) any later version.

Modules

asymmetric
Asymmetric Encryption Module.

symmetric
Symmetric encryption module.