Module proj_crypto::asymmetric

Asymmetric encryption module.

See submodules for documentation.

Unfortunately key_exchange and sign have to use a different representation of the keypair because libsodium uses incompatible representations of the public key between signatures and key exchanges (public interface to curve25519_mult_base). The types which are part of this module are appropriate for usage with key_exchange and key_id.

Modules

commitments
Implementation of Pedersen's Commitment scheme http://download.springer.com/static/pdf/357/chp%253A10.1007%252F3-540-46766-1_9.pdf

key_exchange
Key exchange module

key_id

sign
Public Key Signatures

Constants

PUBLIC_KEY_BYTES
The number of bytes in a PublicKey

SECRET_KEY_BYTES
The number of bytes in a SecretKey

Functions

public_key_from_slice
create a public key from a slice

secret_key_from_slice
create a secret key from a slice

Type Definitions

PublicKey
Public Key - just an alias. Implements drop() so the memory will be wiped when it goes out of scope

SecretKey
Secret Key - just an alias. Implements drop() so the memory will be wiped when it goes out of scope