Struct proj_crypto::asymmetric::LongTermKeys

pub struct LongTermKeys {
    pub my_public_key: PublicKey,
    pub my_secret_key: SecretKey,
    pub their_public_key: PublicKey,
}

Stores long term keys (e.g. from a certificate authority) The secret key is safely erased from memory when this goes out of scope

Fields
my_public_key: PublicKey

This party's long term public key

my_secret_key: SecretKey

This party's long term secret key. Secret key implements drop so that it is wiped when it goes out of scope.

their_public_key: PublicKey

The public key of the party we are communicating with

Methods

impl LongTermKeys

fn server_first(&self, device_ephemeral_public: &PublicKey, message_number: u16) -> (Vec<u8>, SessionKeys, [u8; 32], Vec<u8>)

the first message from the server. This comes after receiving the first message from the device.

Returns (the random challenge, the session keys, the authentication tag to send to the device, the plaintext to send to the device)

fn device_verify_server_msg(&self, ephemeral_pk: &PublicKey, ephemeral_sk: &SecretKey, msg: &[u8], server_message_n: u16, auth_tag: &[u8]) -> bool

for verifying the first message sent by the server

fn device_second(&self, server_ephemeral_pk: &PublicKey, random_challenge: &[u8], ephemeral_pk: &PublicKey, ephemeral_sk: &SecretKey, message_n: u16) -> (SessionKeys, Vec<u8>)

The second message sent by the device. As far as the device is concerned, the setup is complete after sending this.

Returns (session keys, ciphertext to send to the server)