Module proj_crypto::symmetric::chacha20hmacsha512256

Module containing an in implementation of AuthenticatedEncryptorDecryptor. This module provides an implementation of AuthenticatedEncryptorDecryptor.

Cryptography

Authenticates first with HMAC-SHA512 (only the first 256 bytes are used). This was chosen as it is the default authentication mechanism in sodiumoxide. Then we encrypt using ChaCha20. ChaCha20 over the sodiumoxide default (xsalsa20) because I will not be using a random nonse and chacha is more resistant to crypt analysis (see it's introductory paper). The key is used directly. You most likely want to hash it before using it here. You may need to also hash the nonce before using it here.

Example

use sodiumoxide::crypto::stream::chacha20;
use sodiumoxide::crypto::auth::hmacsha512256;
use std::str;
 
sodiumoxide::init();
let e_k = chacha20::gen_key();       // encryption key
let a_k = hmacsha512256::gen_key();  // authentication key
let nonce = chacha20::gen_nonce();   // nonce
let dut = ChaCha20HmacSha512256::new(e_k, a_k, nonce);

let message = "hello world!";

let ciphertext = dut.authenticate_and_encrypt(message.as_bytes());

let transmitted_message = dut.decrypt_and_authenticate(&ciphertext).unwrap();

assert_eq!(message, str::from_utf8(&transmitted_message).unwrap());

Reexports

pub use super::AuthenticatedEncryptorDecryptor;

Structs

ChaCha20HmacSha512256
Struct for storing the state of the symmetric encryption and authentication system. You do not need to worry about destroying these data properly as this is done within sodiumoxide whenever it's types go out of scope.