Struct proj_crypto::symmetric::chacha20hmacsha512256::ChaCha20HmacSha512256

pub struct ChaCha20HmacSha512256 {
    // some fields omitted
}

Struct for storing the state of the symmetric encryption and authentication system. You do not need to worry about destroying these data properly as this is done within sodiumoxide whenever it's types go out of scope.

Methods

impl ChaCha20HmacSha512256

fn new(encryption_key: Key, authentication_key: Key, nonce: Nonce) -> ChaCha20HmacSha512256

Constructor for ChaCha20HmacSha512256.

Example

use sodiumoxide::crypto::stream::chacha20;
use sodiumoxide::crypto::auth::hmacsha512256;
 
sodiumoxide::init();
ChaCha20HmacSha512256::new( chacha20::gen_key(), hmacsha512256::gen_key(), chacha20::gen_nonce() );