Trait proj_crypto::symmetric::AuthenticatedEncryptorDecryptor

pub trait AuthenticatedEncryptorDecryptor {
    fn authenticate_and_encrypt(&self, message: &[u8]) -> Vec<u8>;
    fn decrypt_and_authenticate(&self, ciphertext: &[u8]) -> Option<Vec<u8>>;
}

Trait representing something that does symmetric authenticated encryption.

Required Methods

fn authenticate_and_encrypt(&self, message: &[u8]) -> Vec<u8>

Authenticates the message and then encrypts the message and the authentication token. The result is returned as a vector.

fn decrypt_and_authenticate(&self, ciphertext: &[u8]) -> Option<Vec<u8>>

Decrypts the cipher text and then attempts to authenticate it.