Module proj_crypto::asymmetric

Asymmetric Encryption Module.

This module performs the cryptography for the asymmetric key exchange. It is pretty much the same as is used in Signal, explained here https://whispersystems.org/blog/advanced-ratcheting/.

Basically the encryption shared secret is derived from ephemeral key pairs and authentication keys are the sender's long-term key pair exchanged with the receiver's ephemeral key pair. This is faster than signing and I think it is rather elegant too.

Key exchange is implemented as suggested here https://download.libsodium.org/doc/advanced/scalar_multiplication.html

The Protocol

Device Message 0

generate ephemeral keypair

send ephemeral key to the server

Server Message 0

Generate ephemeral keypair

Compute session keys

Pick a random challenge number

Send ephemeral public key and r to the client, plaintext authentication (as the client does not yet have the encryption key)

Device Message 1

Check auth

Compute session keys

Send r to server, encrypted and authenticated. This authenticates the ephemeral public key we sent in message 0

Server

Decrypt and authenticate and check the challenge response

An important note:

Authentication session keys are symmetric therefore either party can impersonate the other. In an interactive setting this is not a problem because the keys are fixed to only this pair and the other side would not be expecting to receive a message authenticated using their key. However, if Bob decided to publish all his key material he could fabricate messages which look to a third party as though they are sent by Alice. This was intentional in the design of Signal's key exchange because it gives both parties plausible deniability.

Example

use std::str;
 
sodiumoxide::init();

// device long keypair
let (d_pk, d_sk) = gen_keypair();

// server long keypair
let (s_pk, s_sk) = gen_keypair();

let device = LongTermKeys {
    my_public_key: d_pk.clone(),
    my_secret_key: d_sk.clone(),
    their_public_key: s_pk.clone(),
};

let server = LongTermKeys {
    my_public_key: s_pk.clone(),
    my_secret_key: s_sk,
    their_public_key: d_pk.clone(),
};

// First message from the device
let (d_pk_session, d_sk_session) = device_first();

// Message from the server
let (challenge, server_session_keys, auth_tag, plaintext) = server.server_first(&d_pk_session, 0);

// The Device should verify this message
assert!(device.device_verify_server_msg(&d_pk_session, &d_sk_session, &plaintext, 0, &auth_tag));

// the device should now deconstruct the server message to get the public key and challenge (this is bad code so that it is short for demo purposes. Don't copy it)
let (server_pub_key_bytes, challenge_recvd) = plaintext.split_at(PUBLIC_KEY_BYTES);
let server_pub_key = public_key_from_slice(&server_pub_key_bytes).unwrap();

// Challenge response from the device
let (device_session_keys, ciphertext) = device.device_second(&server_pub_key, challenge_recvd, &d_pk_session, &d_sk_session, 1);

// server verifying the client's challenge response
assert!(server_verify_response(&server_session_keys, &ciphertext, 1, &challenge));

// we cannot access the keys directly so let's just check that we can read messages sent by each-other
let message = "hello world!".as_bytes();
let msg_from_server = server_session_keys.from_server.authenticated_encryption(&message, 1);
let msg_from_device = device_session_keys.from_device.authenticated_encryption(&message, 2);

let msg_recvd_server = server_session_keys.from_device.authenticated_decryption(&msg_from_device, 2).unwrap();
let msg_recvd_device = device_session_keys.from_server.authenticated_decryption(&msg_from_server, 1).unwrap();

assert_eq!(msg_recvd_server, message);
assert_eq!(msg_recvd_device, message);

Structs

LongTermKeys
Stores long term keys (e.g. from a certificate authority) The secret key is safely erased from memory when this goes out of scope

SessionKeys
stores session keys

Constants

CHALLENGE_BYTES
The length of the challenge sent by the server to the client

PUBLIC_KEY_BYTES
The number of bytes in a PublicKey

SECRET_KEY_BYTES
The number of bytes in a SecretKey

Functions

device_first
The first message from the device. This initiates the exchange.

gen_keypair
Generate an asymmetric key pair.

public_key_from_slice
create a public key from a slice

secret_key_from_slice
create a secret key from a slice

server_verify_response
For the server to verify the challenge response from the client.

Type Definitions

PublicKey
Public Key - just an alias. Implements drop() so the memory will be wiped when it goes out of scope

SecretKey
Secret Key - just an alias. Implements drop() so the memory will be wiped when it goes out of scope